Replace AbstractCompactionStrategy inheritance with delegation
