+++ java/compaction-job-creation/src/main/java/sleeper/compaction/strategy/impl/CompactionFactory.java
@@ -1,0 +1,109 @@
+/*
+ * Copyright 2022 Crown Copyright
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package sleeper.compaction.strategy.impl;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import sleeper.compaction.job.CompactionJob;
+import sleeper.compaction.job.creation.CreateCompactionJob;
+import sleeper.configuration.properties.InstanceProperties;
+import sleeper.configuration.properties.table.TableProperties;
+import sleeper.core.schema.Schema;
+import sleeper.statestore.FileInfo;
+
+import java.util.List;
+import java.util.UUID;
+
+import static sleeper.configuration.properties.UserDefinedInstanceProperty.FILE_SYSTEM;
+import static sleeper.configuration.properties.table.TableProperty.DATA_BUCKET;
+import static sleeper.configuration.properties.table.TableProperty.ITERATOR_CLASS_NAME;
+import static sleeper.configuration.properties.table.TableProperty.ITERATOR_CONFIG;
+import static sleeper.configuration.properties.table.TableProperty.TABLE_NAME;
+
+public class CompactionFactory {
+    private static final Logger LOGGER = LoggerFactory.getLogger(CompactionFactory.class);
+
+    private Schema schema;
+    private String tableName;
+    private String tableBucket;
+    private String fs;
+    private String iteratorClassName;
+    private String iteratorConfig;
+
+    public void init(InstanceProperties instanceProperties, TableProperties tableProperties) {
+        schema = tableProperties.getSchema();
+        tableName = tableProperties.get(TABLE_NAME);
+        tableBucket = tableProperties.get(DATA_BUCKET);
+        fs = instanceProperties.get(FILE_SYSTEM);
+        iteratorClassName = tableProperties.get(ITERATOR_CLASS_NAME);
+        iteratorConfig = tableProperties.get(ITERATOR_CONFIG);
+        LOGGER.info("Initialised CompactionFactory with table name {}, bucket for table data {}, fs {}",
+                tableName, tableBucket, fs);
+    }
+
+    public CompactionJob createSplittingCompactionJob(
+            List<FileInfo> files, String partition, String leftPartitionId, String rightPartitionId,
+            Object splitPoint, int dimension) {
+        return createSplittingCompaction(
+                files, partition, leftPartitionId, rightPartitionId,
+                splitPoint, dimension)
+                .buildJob();
+    }
+
+    public CreateCompactionJob createSplittingCompaction(
+            List<FileInfo> files, String partition, String leftPartitionId, String rightPartitionId,
+            Object splitPoint, int dimension) {
+
+        String jobId = UUID.randomUUID().toString();
+        String leftOutputFile = fs + tableBucket + "/partition_" + leftPartitionId + "/" + jobId + ".parquet";
+        String rightOutputFile = fs + tableBucket + "/partition_" + rightPartitionId + "/" + jobId + ".parquet";
+        CreateCompactionJob create = createJob().jobId(jobId).partitionId(partition).inputFiles(files)
+                .splitting(split -> split.leftPartitionId(leftPartitionId).rightPartitionId(rightPartitionId)
+                        .leftOutputFile(leftOutputFile).rightOutputFile(rightOutputFile)
+                        .splitPoint(splitPoint).dimension(dimension)
+                        .build())
+                .build();
+
+        LOGGER.info("Created compaction job of id {} to compact and split {} files in partition {}, into partitions {} and {}, to output files {}, {}",
+                jobId, files.size(), partition, leftPartitionId, rightPartitionId, leftOutputFile, rightOutputFile);
+
+        return create;
+    }
+
+    public CompactionJob createCompactionJob(
+            List<FileInfo> files, String partition) {
+        return createCompaction(files, partition).buildJob();
+    }
+
+    public CreateCompactionJob createCompaction(
+            List<FileInfo> files, String partition) {
+
+        String jobId = UUID.randomUUID().toString();
+        String outputFile = fs + tableBucket + "/partition_" + partition + "/" + jobId + ".parquet";
+        CreateCompactionJob create = createJob().jobId(jobId).partitionId(partition)
+                .inputFiles(files).outputFilePath(outputFile).build();
+
+        LOGGER.info("Created compaction job of id {} to compact and split {} files in partition {} to output file {}",
+                jobId, files.size(), partition, outputFile);
+
+        return create;
+    }
+
+    public CreateCompactionJob.Builder createJob() {
+        return CreateCompactionJob.builder().tableName(tableName)
+                .iteratorClassName(iteratorClassName).iteratorConfig(iteratorConfig);
+    }
+}
