+++ java/compaction-job-creation/src/main/java/sleeper/compaction/strategy/impl/LeafPartitionCompactionStrategy.java
@@ -15,17 +15,17 @@
  */
 package sleeper.compaction.strategy.impl;
 
-import sleeper.compaction.strategy.CompactionStrategy;
+import sleeper.compaction.job.CompactionJob;
+import sleeper.configuration.properties.InstanceProperties;
+import sleeper.configuration.properties.table.TableProperties;
+import sleeper.core.partition.Partition;
+import sleeper.statestore.FileInfo;
 
-/**
- * A simple {@link CompactionStrategy} that lists the active files for a partition in increasing order of the number
- * of records they contain, and iterates through this list creating compaction jobs with at most
- * maximumNumberOfFilesToCompact files in each.
- */
-public class BasicCompactionStrategy extends DelegatingCompactionStrategy {
+import java.util.List;
 
-    public BasicCompactionStrategy() {
-        super(new BasicLeafStrategy());
-    }
+public interface LeafPartitionCompactionStrategy {
 
+    void init(InstanceProperties instanceProperties, TableProperties tableProperties, CompactionFactory factory);
+
+    List<CompactionJob> createJobsForLeafPartition(Partition partition, List<FileInfo> activeFilesWithNoJobId);
 }
